New version of gmock included in thirdparty
The CMakeLists are modified to be used in both windows VS and linux. They automatically install the generated libraries to the lib corresponding lib folders.
GTest is now within gmock, and as they say in the installation instructions, we use the version included with gmock.
The CMakeLists of the hybridautomaton project are changed so that we only look for gtest and gmock in the thirdparty folder, so that the user can have another version installed in his computer