_initializeGoal is now called _updateAbsoluteGoal (in Controller and all the derived classes that reimplement it)
Controller::initialize() was before empty but now it calls controller::updateGoal(). Controller::initialize() should not be called directly but indirectly from the derived classes that reimplement derivedControllers::initialize()
Controller::updateGoal() is a new method that converts (if necessary) the goal from relative to absolute by calling the method Controller::relativeToAbsolute. Then it passes the outcome to Controller::_updateAbsoluteGoal

The derived classes do not need to reimplement Controller::updateGoal (it actually does not need to be virtual), but they usually reimplement:
- Controller::initialize() -> Does a lot of Rlab initialization and at the end calls the base class implementation of Controller::initialize() that just calls updateGoal
- Controller::_updateAbsoluteGoal -> Also Rlab things to set correctly the goal
- Controller::relativeGoalToAbsolute -> Depending on the type of controller (and therefore, of goal) the conversion from relative to absolute is different